+++ src/bin/print.rs
@@ -1,0 +1,7 @@
+extern crate peroxide;
+use peroxide::*;
+
+fn main() {
+    let a = zeros(100, 100);
+    a.print();
+}