+++ src/lib.rs
@@ -69,3 +69,6 @@ pub use numerical::runge_kutta::*;
 
 #[allow(unused_imports)]
 pub use util::api::*;
+
+#[allow(unused_imports)]
+pub use numerical::ode::*;